Mild reorg of recent changes.
